Use double for GJK/EPA algorithm

+++ physicsEngine/src/collision/narrowphase/algorithm/ConvexHullConvexHullCollisionAlgorithm.cpp
@@ -25,39 +25,39 @@ namespace urchin
 		std::shared_ptr<CollisionConvexObject3D> convexObject2 = object2.getShape().toConvexObject(object2.getShapeWorldTransform());
 
 		//process GJK and EPA hybrid algorithms
-		std::unique_ptr<GJKResult<float>> gjkResultWithoutMargin = gjkAlgorithm.processGJK(*convexObject1.get(), *convexObject2.get(), false);
+		std::unique_ptr<GJKResult<double>> gjkResultWithoutMargin = gjkAlgorithm.processGJK(*convexObject1.get(), *convexObject2.get(), false);
 
 		if(gjkResultWithoutMargin->isValidResult())
 		{
 			if(gjkResultWithoutMargin->isCollide())
 			{ //collision detected on reduced objects (without margins)
-				std::unique_ptr<GJKResult<float>> gjkResultWithMargin = gjkAlgorithm.processGJK(*convexObject1.get(), *convexObject2.get(), true);
+				std::unique_ptr<GJKResult<double>> gjkResultWithMargin = gjkAlgorithm.processGJK(*convexObject1.get(), *convexObject2.get(), true);
 
 				if(gjkResultWithMargin->isValidResult() && gjkResultWithMargin->isCollide())
 				{
-					std::unique_ptr<EPAResult<float>> epaResult = epaAlgorithm.processEPA(*convexObject1.get(), *convexObject2.get(), *gjkResultWithMargin.get());
+					std::unique_ptr<EPAResult<double>> epaResult = epaAlgorithm.processEPA(*convexObject1.get(), *convexObject2.get(), *gjkResultWithMargin.get());
 
 					if(epaResult->isValidResult() && epaResult->isCollide())
 					{ //should be always true except for problems due to float imprecision
-						const Vector3<float> &normalFromObject2 = -epaResult->getNormal();
-						const Point3<float> &pointOnObject2 = epaResult->getContactPointB();
+						const Vector3<double> &normalFromObject2 = (-epaResult->getNormal());
+						const Point3<double> &pointOnObject2 = epaResult->getContactPointB();
 						const float penetrationDepth = -epaResult->getPenetrationDepth();
 
-						addNewContactPoint(normalFromObject2, pointOnObject2, penetrationDepth);
+						addNewContactPoint(normalFromObject2.cast<float>(), pointOnObject2.cast<float>(), (float)penetrationDepth);
 					}
 				}
 			}else
 			{ //collision detected on enlarged objects (with margins) OR no collision detected
-				const Vector3<float> &vectorBA = gjkResultWithoutMargin->getClosestPointB().vector(gjkResultWithoutMargin->getClosestPointA());
+				const Vector3<double> &vectorBA = gjkResultWithoutMargin->getClosestPointB().vector(gjkResultWithoutMargin->getClosestPointA());
 				float vectorBALength = vectorBA.length();
 				float sumMargins = convexObject1->getOuterMargin() + convexObject2->getOuterMargin();
 				if(sumMargins > vectorBALength - getContactBreakingThreshold())
 				{ //collision detected on enlarged objects
-					const Vector3<float> &normalFromObject2 = vectorBA.normalize();
-					const Point3<float> &pointOnObject2 = gjkResultWithoutMargin->getClosestPointB().translate(normalFromObject2 * convexObject2->getOuterMargin());
+					const Vector3<double> &normalFromObject2 = vectorBA.normalize();
+					const Point3<double> &pointOnObject2 = gjkResultWithoutMargin->getClosestPointB().translate(normalFromObject2 * (double)convexObject2->getOuterMargin());
 					const float penetrationDepth = vectorBALength - sumMargins;
 
-					addNewContactPoint(normalFromObject2, pointOnObject2, penetrationDepth);
+					addNewContactPoint(normalFromObject2.cast<float>(), pointOnObject2.cast<float>(), (float)penetrationDepth);
 				}
 			}
 		}
